Remove warnings
Cherry-pick of https://github.com/trinodb/trino/pull/10965/commits/301ff47420b2ab1fada90fd39701fa532935f96e

+++ presto-main/src/main/java/com/facebook/presto/operator/BigintGroupByHash.java
@@ -291,7 +291,7 @@ public class BigintGroupByHash
 
         // An estimate of how much extra memory is needed before we can go ahead and expand the hash table.
         // This includes the new capacity for values, groupIds, and valuesByGroupId as well as the size of the current page
-        preallocatedMemoryInBytes = newCapacity * (long) (Long.BYTES + Integer.BYTES) + calculateMaxFill(newCapacity) * Long.BYTES + currentPageSizeInBytes;
+        preallocatedMemoryInBytes = newCapacity * (long) (Long.BYTES + Integer.BYTES) + (long) calculateMaxFill(newCapacity) * Long.BYTES + currentPageSizeInBytes;
         if (!updateMemory.update()) {
             // reserved memory but has exceeded the limit
             return false;
